pop-up-fixed
in Popup.jsx the field inputs were being stored in a nested "profile" collection inside the root "users' collection for each signed in used.
this would cause conflicts because the edited data in settings and in popup were pointing to two different collections.   to fix this made sure that the pop field inputs were stored in the root collection where the settings fields were being stored