update-api-store to datastore
the api allows a specific number of fetches daily. Every time I would refresh the browser it would fetch brand new meals which led to exceeding the daily quota. to remedy this I created a function that stores a meal with the current day's date, the function the checks if any meals are saved with today's date if there are then it doesn't fetch any more meals.